requirements.txt
tkinter
tkinter.ttk
PIL
requests
pyTwitchAPI
